Reduce memory leakage related to np.polyfit
First of a few areas that I think we can cleanup some variables that are causing some memory leaks.  NOTE: I'm attempting to run this on a 4gb ram raspberry pi.  If it's ok, i'll continue working through the chain of finding what areas are causing memory utilization to spike.